Moved data type generation to it's own script. Fixed vao initialization for opengl render target.
